organize project resources
